Fix adjacent_find usage in `quantile()`
The return value of the function used in adjacent_find is only needed to
halt the search through the vector.  As a result, the returned iterator
has no real use.  Hence, don't bother checking to see if the iterator
points to the end.  It may in fact be necessary to iterate through the
whole vector in order to calculate the desired quantile.